requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.1
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==20.1.0